Feat: Finish Jwt Validation of Ridder and Supply Order

+++ src/purchaseOrder/purchaseOrder.service.ts
@@ -1,5 +1,5 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { eq, sql } from 'drizzle-orm';
+import { desc, and, eq, sql } from 'drizzle-orm';
 import { DRIZZLE } from '../../src/drizzle/drizzle.module';
 import { DrizzleDB } from '../../src/drizzle/types/drizzle';
 import { CreatePurchaseOrderDto } from './dto/create-purchaseOrder.dto';
@@ -11,6 +11,10 @@ import {
   GetAdjacentPurchaseOrdersDto, 
   GetSimilarRoutePurchaseOrdersDto 
 } from './dto/get-purchaseOrder.dto';
+import { PassengerTable } from '../drizzle/schema/passenger.schema';
+import { PassengerInfoTable } from '../drizzle/schema/passengerInfo.schema';
+import { RidderTable } from '../drizzle/schema/ridder.schema';
+import { RidderInfoTable } from '../drizzle/schema/ridderInfo.schema';
 
 @Injectable()
 export class PurchaseOrderService {
@@ -34,7 +38,6 @@ export class PurchaseOrderService {
       isUrgent: createPurchaseOrderDto.isUrgent ?? undefined,
     }).returning({
       id: PurchaseOrderTable.id,
-      creatorId: PurchaseOrderTable.creatorId,
       createdAt: PurchaseOrderTable.createdAt,
       status: PurchaseOrderTable.status,
     });
@@ -43,22 +46,6 @@ export class PurchaseOrderService {
 
 
   /* ================================= Get operations ================================= */
-  async getPurchaseOrderById(id: string) {
-    return await this.db.select({
-      id: PurchaseOrderTable.id,
-      creatorId: PurchaseOrderTable.creatorId,
-      initPrice: PurchaseOrderTable.initPrice,
-      startCord: PurchaseOrderTable.startCord,
-      endCord: PurchaseOrderTable.endCord,
-      createdAt: PurchaseOrderTable.createdAt,
-      updatedAt: PurchaseOrderTable.updatedAt,
-      startAfter: PurchaseOrderTable.startAfter,
-      isUrgent: PurchaseOrderTable.isUrgent,
-      status: PurchaseOrderTable.status,
-    }).from(PurchaseOrderTable)
-      .where(eq(PurchaseOrderTable.id, id));
-  }
-
   async getPurchaseOrdersByCreatorId(
     creatorId: string, 
     limit: number, 
@@ -66,7 +53,6 @@ export class PurchaseOrderService {
   ) {
     return await this.db.select({
       id: PurchaseOrderTable.id,
-      creatorId: PurchaseOrderTable.creatorId,
       initPrice: PurchaseOrderTable.initPrice,
       startCord: PurchaseOrderTable.startCord,
       endCord: PurchaseOrderTable.endCord,
@@ -77,15 +63,16 @@ export class PurchaseOrderService {
       status: PurchaseOrderTable.status,
     }).from(PurchaseOrderTable)
       .where(eq(PurchaseOrderTable.creatorId, creatorId))
-      .orderBy(PurchaseOrderTable.updatedAt)
+      .orderBy(desc(PurchaseOrderTable.updatedAt))
       .limit(limit)
       .offset(offset);
   }
 
-  async getPurchaseOrders(limit: number, offset: number) {
+  async getPurchaseOrderById(id: string) {
     return await this.db.select({
       id: PurchaseOrderTable.id,
-      creatorId: PurchaseOrderTable.creatorId,
+      creatorName: PassengerTable.userName,
+      avatorUrl: PassengerInfoTable.avatorUrl,
       initPrice: PurchaseOrderTable.initPrice,
       startCord: PurchaseOrderTable.startCord,
       endCord: PurchaseOrderTable.endCord,
@@ -95,19 +82,68 @@ export class PurchaseOrderService {
       isUrgent: PurchaseOrderTable.isUrgent,
       status: PurchaseOrderTable.status,
     }).from(PurchaseOrderTable)
-      .orderBy(PurchaseOrderTable.updatedAt)
-      .limit(limit)
-      .offset(offset)
+      .where(eq(PurchaseOrderTable.id, id)) // basically, should only get one result
+      .leftJoin(PassengerTable, eq(PurchaseOrderTable.creatorId, PassengerTable.id))
+      .leftJoin(PassengerInfoTable, eq(PassengerTable.id, PassengerInfoTable.userId))
+      .limit(1);
   }
 
-  async getCurAdjacentPurchaseOrders(
+  /* ================= Search operations ================= */
+  async searchPurchaseOrderByCreatorName(
+    creatorName: string,
+    limit: number,
+    offset: number,
+  ) {
+    return await this.db.select({
+      id: PurchaseOrderTable.id,
+      creatorName: RidderTable.userName,
+      avatorUrl: RidderInfoTable.avatorUrl,
+      initPrice: PurchaseOrderTable.initPrice,
+      startCord: PurchaseOrderTable.startCord,
+      endCord: PurchaseOrderTable.endCord,
+      createdAt: PurchaseOrderTable.createdAt,
+      updatedAt: PurchaseOrderTable.updatedAt,
+      startAfter: PurchaseOrderTable.startAfter,
+      isUrgent: PurchaseOrderTable.isUrgent,
+      status: PurchaseOrderTable.status,
+    }).from(PurchaseOrderTable)
+      .leftJoin(PassengerTable, eq(PurchaseOrderTable.creatorId, PassengerTable.id))
+      .where(eq(PassengerTable.userName, creatorName))
+      .leftJoin(PassengerInfoTable, eq(PassengerTable.id, PassengerInfoTable.userId))
+      .orderBy(PurchaseOrderTable.updatedAt)
+      .limit(limit)
+      .offset(offset);
+  }
+
+  async searchPaginationPurchaseOrders(limit: number, offset: number) {
+    return await this.db.select({
+      id: PurchaseOrderTable.id,
+      creatorName: PassengerTable.userName,
+      avatorUrl: PassengerInfoTable.avatorUrl,
+      initPrice: PurchaseOrderTable.initPrice,
+      startCord: PurchaseOrderTable.startCord,
+      endCord: PurchaseOrderTable.endCord,
+      createdAt: PurchaseOrderTable.createdAt,
+      updatedAt: PurchaseOrderTable.updatedAt,
+      startAfter: PurchaseOrderTable.startAfter,
+      isUrgent: PurchaseOrderTable.isUrgent,
+      status: PurchaseOrderTable.status,
+    }).from(PurchaseOrderTable)
+      .leftJoin(PassengerTable, eq(PurchaseOrderTable.creatorId, PassengerTable.id))
+      .leftJoin(PassengerInfoTable, eq(PassengerTable.id, PassengerInfoTable.userId))
+      .limit(limit)
+      .offset(offset);
+  }
+
+  async searchCurAdjacentPurchaseOrders(
     limit: number,
     offset: number,
     getAdjacentPurchaseOrdersDto: GetAdjacentPurchaseOrdersDto
   ) {
     return await this.db.select({
       id: PurchaseOrderTable.id,
-      creatorId: PurchaseOrderTable.creatorId,
+      creatorName: PassengerTable.userName,
+      avatorUrl: PassengerInfoTable.avatorUrl,
       initPrice: PurchaseOrderTable.initPrice,
       startCord: PurchaseOrderTable.startCord,
       endCord: PurchaseOrderTable.endCord,
@@ -121,6 +157,8 @@ export class PurchaseOrderService {
         ST_SetSRID(ST_MakePoint(${getAdjacentPurchaseOrdersDto.cordLongitude}, ${getAdjacentPurchaseOrdersDto.cordLatitude}), 4326)
       )`
     }).from(PurchaseOrderTable)
+      .leftJoin(PassengerTable, eq(PurchaseOrderTable.creatorId, PassengerTable.id))
+      .leftJoin(PassengerInfoTable, eq(PassengerTable.id, PassengerInfoTable.userId))
       .orderBy(sql`ST_Distance(
         ${PurchaseOrderTable.startCord},
         ST_SetSRID(ST_MakePoint(${getAdjacentPurchaseOrdersDto.cordLongitude}, ${getAdjacentPurchaseOrdersDto.cordLatitude}), 4326)
@@ -129,14 +167,15 @@ export class PurchaseOrderService {
       .offset(offset);
   }
 
-  async getDestAdjacentPurchaseOrders(
+  async searchDestAdjacentPurchaseOrders(
     limit: number,
     offset: number,
     getAdjacentPurchaseOrdersDto: GetAdjacentPurchaseOrdersDto
   ) {
     return await this.db.select({
       id: PurchaseOrderTable.id,
-      creatorId: PurchaseOrderTable.creatorId,
+      creatorName: PassengerTable.userName,
+      avatorUrl: PassengerInfoTable.avatorUrl,
       initPrice: PurchaseOrderTable.initPrice,
       startCord: PurchaseOrderTable.startCord,
       endCord: PurchaseOrderTable.endCord,
@@ -150,6 +189,8 @@ export class PurchaseOrderService {
         ST_SetSRID(ST_MakePoint(${getAdjacentPurchaseOrdersDto.cordLongitude}, ${getAdjacentPurchaseOrdersDto.cordLatitude}), 4326)
       )`
     }).from(PurchaseOrderTable)
+      .leftJoin(PassengerTable, eq(PurchaseOrderTable.creatorId, PassengerTable.id))
+      .leftJoin(PassengerInfoTable, eq(PassengerTable.id, PassengerInfoTable.userId))
       .orderBy(sql`ST_Distance(
         ${PurchaseOrderTable.endCord},
         ST_SetSRID(ST_MakePoint(${getAdjacentPurchaseOrdersDto.cordLongitude}, ${getAdjacentPurchaseOrdersDto.cordLatitude}), 4326)
@@ -158,14 +199,15 @@ export class PurchaseOrderService {
       .offset(offset);
   }
 
-  async getSimilarRoutePurchaseOrders(
+  async searchSimilarRoutePurchaseOrders(
     limit: number,
     offset: number,
     getSimilarRoutePurchaseOrdersDto: GetSimilarRoutePurchaseOrdersDto
   ) {
     return await this.db.select({
       id: PurchaseOrderTable.id,
-      creatorId: PurchaseOrderTable.creatorId,
+      creatorName: PassengerTable.userName,
+      avatorUrl: PassengerInfoTable.avatorUrl,
       initPrice: PurchaseOrderTable.initPrice,
       startCord: PurchaseOrderTable.startCord,
       endCord: PurchaseOrderTable.endCord,
@@ -193,6 +235,8 @@ export class PurchaseOrderService {
           )
       `,
     }).from(PurchaseOrderTable)
+      .leftJoin(PassengerTable, eq(PurchaseOrderTable.creatorId, PassengerTable.id))
+      .leftJoin(PassengerInfoTable, eq(PassengerTable.id, PassengerInfoTable.userId))
       .orderBy(sql`
           ST_Distance(
             ${PurchaseOrderTable.startCord},
@@ -210,15 +254,25 @@ export class PurchaseOrderService {
             ${PurchaseOrderTable.startCord},
             ${PurchaseOrderTable.endCord}
           )
-      `).limit(limit)
-        .offset(offset);
+      `)
+      .limit(limit)
+      .offset(offset);
   }
+  /* ================= Search operations ================= */
+
   /* ================================= Get operations ================================= */
 
 
   /* ================================= Update operations ================================= */
-  async updatePurchaseOrderById(id: string, updatePurchaseOrderDto: UpdatePurchaseOrderDto) {
-    // we should code like the below form for update method
+  async updatePurchaseOrderById(
+    id: string, // id of the order, for finding the specific order
+    creatorId: string,  // id of the passenger who want to update this order, for checking the validation
+    updatePurchaseOrderDto: UpdatePurchaseOrderDto
+  ) {
+    // validate if the creator of this order(represented by orderID)
+    // is the same as current passenger, since a user can only update the order created by himself
+    // we do this by also include equality of the given creatorId and creatorId of that specific order
+
     const newStartCord: point | undefined = 
       (updatePurchaseOrderDto.startCordLongitude !== undefined 
         && updatePurchaseOrderDto.startCordLatitude !== undefined)
@@ -240,10 +294,9 @@ export class PurchaseOrderService {
       startAfter: updatePurchaseOrderDto.startAfter,
       isUrgent: updatePurchaseOrderDto.isUrgent,
       status: updatePurchaseOrderDto.status,
-    }).where(eq(PurchaseOrderTable.id, id))
+    }).where(and(eq(PurchaseOrderTable.id, id), eq(PurchaseOrderTable.creatorId, creatorId)))
       .returning({
         id: PurchaseOrderTable.id,
-        creatorId: PurchaseOrderTable.creatorId,
         updatedAt: PurchaseOrderTable.updatedAt,
         status: PurchaseOrderTable.status,
       });
@@ -252,9 +305,15 @@ export class PurchaseOrderService {
 
 
   /* ================================= Delete operations ================================= */
-  async deletePurchaseOrderById(id: string) {
+  async deletePurchaseOrderById(id: string, creatorId: string) {
+    // do the same check as update, since the passenger can only delete the order created by himself
     return await this.db.delete(PurchaseOrderTable)
-      .where(eq(PurchaseOrderTable.id, id));
+      .where(and(eq(PurchaseOrderTable.id, id), eq(PurchaseOrderTable.creatorId, creatorId)))
+      .returning({
+        id: PurchaseOrderTable.id,
+        deletedAt: PurchaseOrderTable.updatedAt,
+        status: PurchaseOrderTable.status,
+      });
   }
   /* ================================= Delete operations ================================= */
